Move the kernels used in the detrector volume class to their own file.
